requirements.txt
mysql-connector-python
streamlit
streamlit_option_menu
